@@ -1,70 +1,42 @@
 {
-  "name": "weathery-client",
-  "version": "0.1.0",
+  "name": "weathery-vite",
   "private": true,
-  "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.32",
-    "@fortawesome/free-brands-svg-icons": "^5.15.1",
-    "@fortawesome/free-solid-svg-icons": "^5.15.1",
-    "@fortawesome/react-fontawesome": "^0.1.11",
-    "@fullhuman/postcss-purgecss": "^2.3.0",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "@types/classnames": "^2.2.10",
-    "@types/jest": "^24.9.1",
-    "@types/jsonwebtoken": "^8.5.0",
-    "@types/node": "^12.12.64",
-    "@types/react": "^16.9.51",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-facebook-login": "^4.1.2",
-    "@types/react-router-dom": "^5.1.5",
-    "@types/react-select": "^3.0.26",
-    "@types/yup": "^0.29.7",
-    "autoprefixer": "^9.8.6",
-    "axios": "^0.21.1",
-    "classnames": "^2.2.6",
-    "concurrently": "^5.3.0",
-    "framer-motion": "^2.9.4",
-    "jsonwebtoken": "^8.5.1",
-    "moment": "^2.29.0",
-    "node-sass": "^4.14.1",
-    "postcss-cli": "^7.1.2",
-    "query-string": "^6.13.5",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-facebook-login": "^4.1.1",
-    "react-fluent-form": "^1.5.3",
-    "react-google-login": "^5.1.21",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^3.4.3",
-    "react-select": "^3.1.1",
-    "react-toastify": "^6.0.9",
-    "tailwindcss": "^1.8.10",
-    "typescript": "^3.9.7",
-    "yup": "^0.29.3"
-  },
+  "version": "0.0.0",
   "scripts": {
-    "start": "concurrently \"npm run watch:css\" \"react-scripts start\"",
-    "build": "npm run build:css && react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "build:css": "postcss src/assets/styles/tailwind.css -o src/assets/styles/main.css",
-    "watch:css": "postcss src/assets/styles/tailwind.css -o src/assets/styles/main.css -w"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview"
   },
-  "eslintConfig": {
-    "extends": "react-app"
+  "dependencies": {
+    "@types/react-facebook-login": "^4.1.4",
+    "axios": "^0.26.1",
+    "date-fns": "^2.28.0",
+    "framer-motion": "^6.2.8",
+    "jwt-decode": "^3.1.2",
+    "query-string": "^7.1.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-facebook-login": "^4.1.1",
+    "react-fluent-form": "^1.6.0",
+    "react-google-login": "^5.2.2",
+    "react-icons": "^4.3.1",
+    "react-router-dom": "^6.2.2",
+    "react-select": "^5.2.2",
+    "react-toastify": "^8.2.0",
+    "sass": "^1.49.9",
+    "yup": "^0.32.11"
   },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "devDependencies": {
+    "@types/node": "^17.0.23",
+    "@types/react": "^17.0.33",
+    "@types/react-dom": "^17.0.10",
+    "@vitejs/plugin-react": "^1.0.7",
+    "autoprefixer": "^10.4.4",
+    "postcss": "^8.4.12",
+    "tailwindcss": "^3.0.23",
+    "typescript": "^4.5.4",
+    "vite": "^2.8.0",
+    "vite-plugin-pwa": "^0.11.13",
+    "workbox-window": "^6.5.2"
   }
 }
