@@ -1,7 +1,8 @@
 import { createForm, field } from "react-fluent-form";
 import * as yup from "yup";
 
-import { City } from "../models/CitySearchResponse";
+
+import { City } from "../api/models/CitySearchResponse";
 import { Nullable } from "../utils/Nullable";
 
 interface CitySearchForm {
@@ -9,7 +10,7 @@ interface CitySearchForm {
 }
 
 export const citySearchFormConfig = createForm<CitySearchForm>()({
-  city: field.singleSelect<City>(null)
+  city: field.singleSelect<City>(null),
 }).withValidation({
-  city: yup.object().required()
+  city: yup.object().required(),
 });
