@@ -12,11 +12,11 @@ export const signupFormConfig = createForm<SignupForm>()({
   email: field.email(),
   password: field.password(),
   passwordConfirm: field.password(),
-  time: field.raw(new Date())
+  time: field.raw(new Date()),
 }).withValidation({
   password: (value: string, values: SignupForm) => {
     if (value !== values.passwordConfirm) {
       return "Password should match Password Confirm";
     }
-  }
+  },
 });
