@@ -1,9 +1,9 @@
 import { motion } from "framer-motion";
-import React, { useContext } from "react";
 import { useFluentForm } from "react-fluent-form";
 
+
 import { forgotPasswordFormConfig } from "../../form-config/ForgotPasswordFormConfig";
-import { AuthServiceContext } from "../../service/auth/AuthServiceContext";
+import { useAuthService } from "../../service/auth/useAuthService";
 import { scaleVariation } from "../common/variants/framerVariants";
 
 interface ForgotPasswordModalProps {
@@ -11,13 +11,13 @@ interface ForgotPasswordModalProps {
 }
 
 export const ForgotPasswordModal: React.FC<ForgotPasswordModalProps> = ({
-  closeFunction
+  closeFunction,
 }) => {
   const { values, fields, handleSubmit, reset } = useFluentForm(
     forgotPasswordFormConfig
   );
 
-  const { forgotPassword, loading } = useContext(AuthServiceContext);
+  const { forgotPassword, loading } = useAuthService();
 
   const handleSubmitSuccess = async () => {
     await forgotPassword(values.email);
