@@ -1,9 +1,9 @@
 import { motion } from "framer-motion";
-import React, { useContext } from "react";
 import { useFluentForm } from "react-fluent-form";
 
+
 import { resetPasswordFormConfig } from "../../form-config/ResetPasswordFormConfig";
-import { AuthServiceContext } from "../../service/auth/AuthServiceContext";
+import { useAuthService } from "../../service/auth/useAuthService";
 import { createErrorToast } from "../../utils/toast/errorToast";
 import { scaleVariation } from "../common/variants/framerVariants";
 
@@ -12,19 +12,18 @@ interface ResetPasswordModalContentProps {
   token: string;
 }
 
-export const ResetPasswordModalContent: React.FC<ResetPasswordModalContentProps> = ({
-  closeFunction,
-  token
-}) => {
+export const ResetPasswordModalContent: React.FC<
+  ResetPasswordModalContentProps
+> = ({ closeFunction, token }) => {
   const {
     values,
     fields,
     handleSubmit,
     reset: resetForm,
-    errors
+    errors,
   } = useFluentForm(resetPasswordFormConfig);
 
-  const { resetPassword, loading } = useContext(AuthServiceContext);
+  const { resetPassword, loading } = useAuthService();
 
   const handleSubmitSuccess = async () => {
     if (await resetPassword(values.password, token)) {
