@@ -17,6 +17,6 @@ const errorToast = (msg: string) => {
     closeOnClick: true,
     pauseOnHover: true,
     draggable: false,
-    progress: undefined
+    progress: undefined,
   });
 };
